+++ FEBuilderGBA/OptionForm.cs
@@ -1101,7 +1101,7 @@ namespace FEBuilderGBA
                     continue;
                 }
 
-                byte[] data = Program.ROM.getBinaryData(start, (uint)need.Length);
+                byte[] data = Program.ROM.getBinaryData(start, need.Length);
                 if (U.memcmp(need, data) != 0)
                 {
                     continue;
